+++ Tests/KeyboardKitTests/Keyboard/KeyboardContextTests.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
-#if os(iOS) || os(tvOS)
 import MockingKit
 import UIKit
 import XCTest
@@ -15,16 +14,14 @@ import XCTest
 
 class KeyboardContextTests: XCTestCase {
 
-    var controller: KeyboardInputViewController!
     var context: KeyboardContext!
     var proxy: MockTextDocumentProxy!
     var traits: MockTraitCollection!
 
     override func setUp() {
-        controller = KeyboardInputViewController()
         proxy = MockTextDocumentProxy()
         traits = MockTraitCollection()
-        context = KeyboardContext(controller: controller, keyboardType: .images)
+        context = KeyboardContext()
         context.traitCollection = traits
         context.textDocumentProxy = proxy
     }
@@ -42,26 +39,53 @@ class KeyboardContextTests: XCTestCase {
         context.hasKeyboardType(type)
     }
 
-
-    func testInitializationAppliesProvidedParams() {
-        context = KeyboardContext(controller: controller, keyboardType: .images)
-
-        XCTAssertEqual(context.keyboardType, .images)
-        XCTAssertEqual(context.locale, .current)
-        XCTAssertEqual(context.locales, [.current])
-
-        XCTAssertEqual(context.deviceType, .current)
+    #if os(iOS) || os(tvOS)
+    func assert(_ context: KeyboardContext, isSyncedWith controller: KeyboardInputViewController) {
         XCTAssertEqual(context.hasDictationKey, controller.hasDictationKey)
         XCTAssertEqual(context.hasFullAccess, controller.hasFullAccess)
         XCTAssertEqual(context.needsInputModeSwitchKey, controller.needsInputModeSwitchKey)
-        XCTAssertNil(context.primaryLanguage)
-        XCTAssertNil(context.textInputMode)
+        XCTAssertEqual(context.primaryLanguage, controller.primaryLanguage)
+        XCTAssertEqual(context.textInputMode, controller.textInputMode)
         eventually {
-            XCTAssertTrue(self.context.textDocumentProxy === self.controller.textDocumentProxy)
-            XCTAssertEqual(self.context.traitCollection, self.controller.traitCollection)
+            XCTAssertTrue(context.textDocumentProxy === controller.textDocumentProxy)
+            XCTAssertEqual(context.traitCollection, controller.traitCollection)
         }
     }
+    #endif
 
+
+    func testInitializerSetsDefaultValues() {
+        XCTAssertEqual(context.deviceType, .current)
+        XCTAssertFalse(context.hasDictationKey)
+        XCTAssertFalse(context.hasFullAccess)
+        XCTAssertNil(context.keyboardDictationReplacement)
+        XCTAssertEqual(context.keyboardType, .alphabetic(.lowercased))
+        XCTAssertEqual(context.locale, .current)
+        XCTAssertEqual(context.locales, [.current])
+        XCTAssertFalse(context.needsInputModeSwitchKey)
+        XCTAssertTrue(context.prefersAutocomplete)
+        XCTAssertNil(context.primaryLanguage)
+        XCTAssertNotNil(context.textDocumentProxy)
+        XCTAssertNil(context.textInputMode)
+        XCTAssertNotNil(context.traitCollection)
+    }
+
+    func testInitializerCanSetLocale() {
+        let id = "sv-SE"
+        let locale = Locale(identifier: id)
+        context = KeyboardContext(locale: locale)
+
+        XCTAssertEqual(context.locale, locale)
+        XCTAssertEqual(context.locales, [locale])
+    }
+
+    #if os(iOS) || os(tvOS)
+    func testInitializerCanSyncWithController() {
+        let controller = KeyboardInputViewController()
+        context = KeyboardContext(controller: controller)
+        assert(context, isSyncedWith: controller)
+    }
+    #endif
 
     func testColorSchemeIsDerivedFromTraitCollection() {
         traits.userInterfaceStyleValue = .light
@@ -70,14 +94,12 @@ class KeyboardContextTests: XCTestCase {
         XCTAssertEqual(context.colorScheme, .dark)
     }
 
-
     func testKeyboardAppearanceIsDerivedFromProxy() {
         proxy.keyboardAppearance = .light
         XCTAssertEqual(context.keyboardAppearance, .light)
         proxy.keyboardAppearance = .dark
         XCTAssertEqual(context.keyboardAppearance, .dark)
     }
-
 
     func testHasKeyboardLocaleReturnsTrueForMatchingType() {
         context.setLocale(.swedish)
@@ -97,7 +119,6 @@ class KeyboardContextTests: XCTestCase {
         XCTAssertFalse(hasKeyboardTypeResult(for: .numeric))
         XCTAssertFalse(hasKeyboardTypeResult(for: .symbolic))
     }
-
 
     func testSelectingNextLocaleSelectsFirstItemIfTheCurrentLocaleIsNotInLocales() {
         context.locale = locale(for: "sv")
@@ -129,22 +150,12 @@ class KeyboardContextTests: XCTestCase {
         XCTAssertEqual(context.locale.identifier, "ca")
     }
 
-
+    #if os(iOS) || os(tvOS)
     func testSyncingContextWithControllerUpdatesSomeProperties() {
-        let context = KeyboardContext(controller: controller, keyboardType: .images)
+        let controller = KeyboardInputViewController()
+        let context = KeyboardContext(controller: controller)
         context.sync(with: controller)
-        XCTAssertEqual(context.hasDictationKey, controller.hasDictationKey)
-        XCTAssertEqual(context.hasFullAccess, controller.hasFullAccess)
-        XCTAssertEqual(context.needsInputModeSwitchKey, controller.needsInputModeSwitchKey)
-        XCTAssertNil(context.primaryLanguage)
-        #if os(iOS) || os(macOS)
-        XCTAssertEqual(context.screenOrientation, controller.screenOrientation)
-        #endif
-        XCTAssertNil(context.textInputMode)
-        eventually {
-            XCTAssertTrue(context.textDocumentProxy === self.controller.textDocumentProxy)
-            XCTAssertEqual(context.traitCollection, self.controller.traitCollection)
-        }
+        assert(context, isSyncedWith: controller)
     }
+    #endif
 }
-#endif